Cleanup occlient package
