+++ pkg/kclient/oc_server.go
@@ -1,0 +1,103 @@
+package kclient
+
+import (
+	"context"
+	"encoding/json"
+	"net"
+	"time"
+
+	"github.com/pkg/errors"
+	"github.com/redhat-developer/odo/pkg/preference"
+	"github.com/redhat-developer/odo/pkg/util"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/version"
+	"k8s.io/klog"
+)
+
+// isServerUp returns true if server is up and running
+// server parameter has to be a valid url
+func isServerUp(server string) bool {
+	// initialising the default timeout, this will be used
+	// when the value is not readable from config
+	ocRequestTimeout := preference.DefaultTimeout * time.Second
+	// checking the value of timeout in config
+	// before proceeding with default timeout
+	cfg, configReadErr := preference.New()
+	if configReadErr != nil {
+		klog.V(3).Info(errors.Wrap(configReadErr, "unable to read config file"))
+	} else {
+		ocRequestTimeout = time.Duration(cfg.GetTimeout()) * time.Second
+	}
+	address, err := util.GetHostWithPort(server)
+	if err != nil {
+		klog.V(3).Infof("Unable to parse url %s (%s)", server, err)
+	}
+	klog.V(3).Infof("Trying to connect to server %s", address)
+	_, connectionError := net.DialTimeout("tcp", address, time.Duration(ocRequestTimeout))
+	if connectionError != nil {
+		klog.V(3).Info(errors.Wrap(connectionError, "unable to connect to server"))
+		return false
+	}
+
+	klog.V(3).Infof("Server %v is up", server)
+	return true
+}
+
+// ServerInfo contains the fields that contain the server's information like
+// address, OpenShift and Kubernetes versions
+type ServerInfo struct {
+	Address           string
+	OpenShiftVersion  string
+	KubernetesVersion string
+}
+
+// GetServerVersion will fetch the Server Host, OpenShift and Kubernetes Version
+// It will be shown on the execution of odo version command
+func (c *Client) GetServerVersion() (*ServerInfo, error) {
+	var info ServerInfo
+
+	// This will fetch the information about Server Address
+	config, err := c.KubeConfig.ClientConfig()
+	if err != nil {
+		return nil, errors.Wrapf(err, "unable to get server's address")
+	}
+	info.Address = config.Host
+
+	// checking if the server is reachable
+	if !isServerUp(config.Host) {
+		return nil, errors.New("Unable to connect to OpenShift cluster, is it down?")
+	}
+
+	// fail fast if user is not connected (same logic as `oc whoami`)
+	_, err = c.userClient.Users().Get(context.TODO(), "~", metav1.GetOptions{})
+	if err != nil {
+		return nil, err
+	}
+
+	// This will fetch the information about OpenShift Version
+	coreGet := c.GetClient().CoreV1().RESTClient().Get()
+	rawOpenShiftVersion, err := coreGet.AbsPath("/version/openshift").Do(context.TODO()).Raw()
+	if err != nil {
+		// when using Minishift (or plain 'oc cluster up' for that matter) with OKD 3.11, the version endpoint is missing...
+		klog.V(3).Infof("Unable to get OpenShift Version - endpoint '/version/openshift' doesn't exist")
+	} else {
+		var openShiftVersion version.Info
+		if e := json.Unmarshal(rawOpenShiftVersion, &openShiftVersion); e != nil {
+			return nil, errors.Wrapf(e, "unable to unmarshal OpenShift version %v", string(rawOpenShiftVersion))
+		}
+		info.OpenShiftVersion = openShiftVersion.GitVersion
+	}
+
+	// This will fetch the information about Kubernetes Version
+	rawKubernetesVersion, err := coreGet.AbsPath("/version").Do(context.TODO()).Raw()
+	if err != nil {
+		return nil, errors.Wrapf(err, "unable to get Kubernetes Version")
+	}
+	var kubernetesVersion version.Info
+	if err := json.Unmarshal(rawKubernetesVersion, &kubernetesVersion); err != nil {
+		return nil, errors.Wrapf(err, "unable to unmarshal Kubernetes Version: %v", string(rawKubernetesVersion))
+	}
+	info.KubernetesVersion = kubernetesVersion.GitVersion
+
+	return &info, nil
+}
