--- pkg/occlient/occlient.go
@@ -1,220 +1,0 @@
-package occlient
-
-import (
-	"context"
-	"encoding/json"
-	"io"
-	"net"
-	"time"
-
-	"github.com/pkg/errors"
-
-	"github.com/redhat-developer/odo/pkg/kclient"
-	"github.com/redhat-developer/odo/pkg/preference"
-	"github.com/redhat-developer/odo/pkg/util"
-
-	// api clientsets
-
-	appsclientset "github.com/openshift/client-go/apps/clientset/versioned/typed/apps/v1"
-	projectclientset "github.com/openshift/client-go/project/clientset/versioned/typed/project/v1"
-	userclientset "github.com/openshift/client-go/user/clientset/versioned/typed/user/v1"
-
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-	"k8s.io/apimachinery/pkg/version"
-	"k8s.io/client-go/tools/clientcmd"
-	"k8s.io/klog"
-)
-
-// CreateArgs is a container of attributes of component create action
-type CreateArgs struct {
-	Name            string
-	SourcePath      string
-	SourceRef       string
-	ImageName       string
-	EnvVars         []string
-	Ports           []string
-	Resources       *corev1.ResourceRequirements
-	ApplicationName string
-	Wait            bool
-	// StorageToBeMounted describes the storage to be created
-	// storagePath is the key of the map, the generatedPVC is the value of the map
-	StorageToBeMounted map[string]*corev1.PersistentVolumeClaim
-	StdOut             io.Writer
-}
-
-//const (
-//	// timeout for waiting for project deletion
-//	waitForProjectDeletionTimeOut = 3 * time.Minute
-//)
-
-// UpdateComponentParams serves the purpose of holding the arguments to a component update request
-type UpdateComponentParams struct {
-	// CommonObjectMeta is the object meta containing the labels and annotations expected for the new deployment
-	CommonObjectMeta metav1.ObjectMeta
-	// ResourceLimits are the cpu and memory constraints to be applied on to the component
-	ResourceLimits corev1.ResourceRequirements
-	// EnvVars to be exposed
-	EnvVars []corev1.EnvVar
-	// StorageToBeMounted describes the storage to be mounted
-	// storagePath is the key of the map, the generatedPVC is the value of the map
-	StorageToBeMounted map[string]*corev1.PersistentVolumeClaim
-	// StorageToBeUnMounted describes the storage to be unmounted
-	// path is the key of the map,storageName is the value of the map
-	StorageToBeUnMounted map[string]string
-}
-
-// errorMsg is the message for user when invalid configuration error occurs
-const errorMsg = `
-Please login to your server: 
-
-odo login https://mycluster.mydomain.com
-`
-
-type Client struct {
-	kubeClient    *kclient.Client
-	appsClient    appsclientset.AppsV1Interface
-	projectClient projectclientset.ProjectV1Interface
-	userClient    userclientset.UserV1Interface
-	KubeConfig    clientcmd.ClientConfig
-	Namespace     string
-}
-
-// New creates a new client
-func New() (*Client, error) {
-	var client Client
-
-	// initialize client-go clients
-	loadingRules := clientcmd.NewDefaultClientConfigLoadingRules()
-	configOverrides := &clientcmd.ConfigOverrides{}
-	client.KubeConfig = clientcmd.NewNonInteractiveDeferredLoadingClientConfig(loadingRules, configOverrides)
-
-	config, err := client.KubeConfig.ClientConfig()
-	if err != nil {
-		return nil, errors.New(err.Error() + errorMsg)
-	}
-
-	client.kubeClient, err = kclient.NewForConfig(client.KubeConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	client.appsClient, err = appsclientset.NewForConfig(config)
-	if err != nil {
-		return nil, err
-	}
-
-	client.projectClient, err = projectclientset.NewForConfig(config)
-	if err != nil {
-		return nil, err
-	}
-
-	client.userClient, err = userclientset.NewForConfig(config)
-	if err != nil {
-		return nil, err
-	}
-
-	client.Namespace, _, err = client.KubeConfig.Namespace()
-	if err != nil {
-		return nil, err
-	}
-
-	return &client, nil
-}
-
-// isServerUp returns true if server is up and running
-// server parameter has to be a valid url
-func isServerUp(server string) bool {
-	// initialising the default timeout, this will be used
-	// when the value is not readable from config
-	ocRequestTimeout := preference.DefaultTimeout * time.Second
-	// checking the value of timeout in config
-	// before proceeding with default timeout
-	cfg, configReadErr := preference.New()
-	if configReadErr != nil {
-		klog.V(3).Info(errors.Wrap(configReadErr, "unable to read config file"))
-	} else {
-		ocRequestTimeout = time.Duration(cfg.GetTimeout()) * time.Second
-	}
-	address, err := util.GetHostWithPort(server)
-	if err != nil {
-		klog.V(3).Infof("Unable to parse url %s (%s)", server, err)
-	}
-	klog.V(3).Infof("Trying to connect to server %s", address)
-	_, connectionError := net.DialTimeout("tcp", address, time.Duration(ocRequestTimeout))
-	if connectionError != nil {
-		klog.V(3).Info(errors.Wrap(connectionError, "unable to connect to server"))
-		return false
-	}
-
-	klog.V(3).Infof("Server %v is up", server)
-	return true
-}
-
-// ServerInfo contains the fields that contain the server's information like
-// address, OpenShift and Kubernetes versions
-type ServerInfo struct {
-	Address           string
-	OpenShiftVersion  string
-	KubernetesVersion string
-}
-
-// GetServerVersion will fetch the Server Host, OpenShift and Kubernetes Version
-// It will be shown on the execution of odo version command
-func (c *Client) GetServerVersion() (*ServerInfo, error) {
-	var info ServerInfo
-
-	// This will fetch the information about Server Address
-	config, err := c.KubeConfig.ClientConfig()
-	if err != nil {
-		return nil, errors.Wrapf(err, "unable to get server's address")
-	}
-	info.Address = config.Host
-
-	// checking if the server is reachable
-	if !isServerUp(config.Host) {
-		return nil, errors.New("Unable to connect to OpenShift cluster, is it down?")
-	}
-
-	// fail fast if user is not connected (same logic as `oc whoami`)
-	_, err = c.userClient.Users().Get(context.TODO(), "~", metav1.GetOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	// This will fetch the information about OpenShift Version
-	coreGet := c.kubeClient.KubeClient.CoreV1().RESTClient().Get()
-	rawOpenShiftVersion, err := coreGet.AbsPath("/version/openshift").Do(context.TODO()).Raw()
-	if err != nil {
-		// when using Minishift (or plain 'oc cluster up' for that matter) with OKD 3.11, the version endpoint is missing...
-		klog.V(3).Infof("Unable to get OpenShift Version - endpoint '/version/openshift' doesn't exist")
-	} else {
-		var openShiftVersion version.Info
-		if e := json.Unmarshal(rawOpenShiftVersion, &openShiftVersion); e != nil {
-			return nil, errors.Wrapf(e, "unable to unmarshal OpenShift version %v", string(rawOpenShiftVersion))
-		}
-		info.OpenShiftVersion = openShiftVersion.GitVersion
-	}
-
-	// This will fetch the information about Kubernetes Version
-	rawKubernetesVersion, err := coreGet.AbsPath("/version").Do(context.TODO()).Raw()
-	if err != nil {
-		return nil, errors.Wrapf(err, "unable to get Kubernetes Version")
-	}
-	var kubernetesVersion version.Info
-	if err := json.Unmarshal(rawKubernetesVersion, &kubernetesVersion); err != nil {
-		return nil, errors.Wrapf(err, "unable to unmarshal Kubernetes Version: %v", string(rawKubernetesVersion))
-	}
-	info.KubernetesVersion = kubernetesVersion.GitVersion
-
-	return &info, nil
-}
-
-func (c *Client) GetKubeClient() *kclient.Client {
-	return c.kubeClient
-}
-
-func (c *Client) SetKubeClient(client *kclient.Client) {
-	c.kubeClient = client
-}
