--- pkg/occlient/occlient_test.go
@@ -1,1 +1,0 @@
-package occlient
